refactor: move intake logic from hugging-face provider
