+++ backend/app/utils/intake_validation.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from typing import Any
+
 
 def _has_answer(value: object) -> bool:
     if value is None:
@@ -26,3 +28,25 @@ def compute_current_index(questions: list[dict], criteria: object) -> int:
         if _has_answer(criteria.get(question_key)):
             count += 1
     return count
+
+
+def _missing_required_fields(
+    merged_criteria: dict[str, Any],
+    required_fields: list[str],
+) -> list[str]:
+    return [key for key in required_fields if key not in merged_criteria]
+
+
+def merge_missing_fields(
+    *,
+    merged_criteria: dict[str, Any],
+    required_fields: list[str],
+    model_missing: list[str],
+) -> list[str]:
+    """Use the model's missing keys when they match real gaps; otherwise criteria-based gaps."""
+    still_missing = _missing_required_fields(merged_criteria, required_fields)
+    from_model = [key for key in model_missing if key in required_fields]
+    if from_model:
+        overlap = [key for key in from_model if key in still_missing]
+        return overlap if overlap else still_missing
+    return still_missing
